+++ src/refactorings/inline/inline-function/find-param-matching-id.ts
@@ -1,4 +1,4 @@
-import * as ast from "../../ast";
+import * as ast from "../../../ast";
 
 export { findParamMatchingId };